+++ packages/components/psammead-caption/src/index.stories.jsx
@@ -1,26 +1,42 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
+import { withKnobs } from '@storybook/addon-knobs';
+import { inputProvider } from '@bbc/psammead-storybook-helpers';
 import InlineLink from '@bbc/psammead-inline-link';
 import VisuallyHiddenText from '@bbc/psammead-visually-hidden-text';
-import { withReadme } from 'storybook-readme';
-import Readme from '../README.md';
+import { latin } from '@bbc/gel-foundations/scripts';
+import notes from '../README.md';
 import Caption from '.';
 
 storiesOf('Caption', module)
-  .addDecorator(withReadme(Readme))
-  .add('default', () => <Caption>This is a caption.</Caption>)
-  .add('with offscreen text', () => (
-    <Caption>
-      <VisuallyHiddenText>Image caption, </VisuallyHiddenText>
-      This is a caption with preceding offscreen text.
-    </Caption>
-  ))
-  .add('containing an inline link', () => (
-    <Caption>
-      {'This is a caption '}
-      <InlineLink href="https://www.bbc.com">
-        containing an inline link
-      </InlineLink>
-      .
-    </Caption>
-  ));
+  .addDecorator(withKnobs)
+  .add(
+    'default',
+    inputProvider(['caption'], captionText => (
+      <Caption script={latin}>{captionText}</Caption>
+    )),
+    { notes, knobs: { escapeHTML: false } },
+  )
+  .add(
+    'with offscreen text',
+    inputProvider(
+      ['visually hidden text', 'caption'],
+      (hiddenText, captionText) => (
+        <Caption>
+          <VisuallyHiddenText>{hiddenText}</VisuallyHiddenText>
+          {captionText}
+        </Caption>
+      ),
+    ),
+    { notes, knobs: { escapeHTML: false } },
+  )
+  .add(
+    'containing an inline link',
+    inputProvider(['inline link', 'caption'], (linkText, captionText) => (
+      <Caption script={latin}>
+        {captionText}
+        <InlineLink href="https://www.bbc.com"> {linkText}</InlineLink>.
+      </Caption>
+    )),
+    { notes, knobs: { escapeHTML: false } },
+  );