+++ packages/components/psammead-paragraph/src/index.stories.jsx
@@ -1,9 +1,17 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { withReadme } from 'storybook-readme';
-import Readme from '../README.md';
+import { latin } from '@bbc/gel-foundations/scripts';
+import { withKnobs } from '@storybook/addon-knobs';
+import { inputProvider } from '@bbc/psammead-storybook-helpers';
+import notes from '../README.md';
 import Paragraph from './index';
 
 storiesOf('Paragraph', module)
-  .addDecorator(withReadme(Readme))
-  .add('default', () => <Paragraph>This is text in a paragraph.</Paragraph>);
+  .addDecorator(withKnobs)
+  .add(
+    'default',
+    inputProvider(['paragraph'], paragraph => (
+      <Paragraph script={latin}>{paragraph}</Paragraph>
+    )),
+    { notes, knobs: { escapeHTML: false } },
+  );