+++ packages/utilities/gel-foundations/prop-types.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/prop-types'); // eslint-disable-line import/no-unresolved