+++ packages/utilities/gel-foundations/src/scripts/sinhalese.js
@@ -1,0 +1,144 @@
+const sinhaleseTypography = {
+  canon: {
+    groupA: {
+      fontSize: '30',
+      lineHeight: '32',
+    },
+    groupB: {
+      fontSize: '38',
+      lineHeight: '40',
+    },
+    groupD: {
+      fontSize: '52',
+      lineHeight: '58',
+    },
+  },
+  trafalgar: {
+    groupA: {
+      fontSize: '24',
+      lineHeight: '24',
+    },
+    groupB: {
+      fontSize: '28',
+      lineHeight: '28',
+    },
+    groupD: {
+      fontSize: '40',
+      lineHeight: '44',
+    },
+  },
+  paragon: {
+    groupA: {
+      fontSize: '24',
+      lineHeight: '24',
+    },
+    groupB: {
+      fontSize: '24',
+      lineHeight: '24',
+    },
+    groupD: {
+      fontSize: '34',
+      lineHeight: '36',
+    },
+  },
+  doublePica: {
+    groupA: {
+      fontSize: '24',
+      lineHeight: '24',
+    },
+    groupB: {
+      fontSize: '24',
+      lineHeight: '24',
+    },
+    groupD: {
+      fontSize: '30',
+      lineHeight: '30',
+    },
+  },
+  greatPrimer: {
+    groupA: {
+      fontSize: '22',
+      lineHeight: '22',
+    },
+    groupB: {
+      fontSize: '24',
+      lineHeight: '24',
+    },
+    groupD: {
+      fontSize: '24',
+      lineHeight: '24',
+    },
+  },
+  bodyCopy: {
+    groupA: {
+      fontSize: '22',
+      lineHeight: '28',
+    },
+    groupB: {
+      fontSize: '22',
+      lineHeight: '28',
+    },
+    groupD: {
+      fontSize: '22',
+      lineHeight: '24',
+    },
+  },
+  pica: {
+    groupA: {
+      fontSize: '22',
+      lineHeight: '24',
+    },
+    groupB: {
+      fontSize: '22',
+      lineHeight: '24',
+    },
+    groupD: {
+      fontSize: '22',
+      lineHeight: '24',
+    },
+  },
+  longPrimer: {
+    groupA: {
+      fontSize: '18',
+      lineHeight: '22',
+    },
+    groupB: {
+      fontSize: '18',
+      lineHeight: '22',
+    },
+    groupD: {
+      fontSize: '18',
+      lineHeight: '22',
+    },
+  },
+  brevier: {
+    groupA: {
+      fontSize: '17',
+      lineHeight: '22',
+    },
+    groupB: {
+      fontSize: '17',
+      lineHeight: '22',
+    },
+    groupD: {
+      fontSize: '17',
+      lineHeight: '22',
+    },
+  },
+  minion: {
+    groupA: {
+      fontSize: '16',
+      lineHeight: '22',
+    },
+    groupB: {
+      fontSize: '16',
+      lineHeight: '22',
+    },
+    groupD: {
+      fontSize: '16',
+      lineHeight: '22',
+    },
+  },
+};
+
+export default sinhaleseTypography;