+++ packages/components/psammead-inline-link/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 | Version | Description |
 |---------|-------------|
+| 0.3.7   | [PR#424](https://github.com/bbc/psammead/pull/424) Add Snyk badge to readme |
+| 0.3.6   | [PR#419](https://github.com/bbc/psammead/pull/419) Add language variants knob to Inline Link stories |
+| 0.3.5   | [PR#407](https://github.com/bbc/psammead/pull/407) Organise dependencies properly |
+| 0.3.4   | [PR#371](https://github.com/bbc/psammead/pull/371) Update colour from Shadow to Ebon |
+| 0.3.3   | [PR#323](https://github.com/bbc/psammead/pull/323) Update storybook badge url |
+| 0.3.2   | [PR#323](https://github.com/BBC/psammead/pull/323) Update readme storybook badge |
 | 0.3.1   | [PR#264](https://github.com/BBC/psammead/pull/319) Resolving package-lock issues. |
 | 0.3.0   | [PR#309](https://github.com/bbc/psammead/pull/309) Update colours to reflect new UX designs. |
 | 0.2.0   | [PR#248](https://github.com/BBC-News/psammead/pull/248) Remove top and buttom padding |