+++ packages/components/psammead-paragraph/CHANGELOG.md
@@ -1,7 +1,17 @@
 # Psammead Paragraph Changelog
 
+<!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 1.0.2   | [PR#424](https://github.com/bbc/psammead/pull/424) Add Snyk badge to readme |
+| 1.0.1   | [PR#420](https://github.com/bbc/psammead/pull/420) Add language variants knob to Paragraph stories |
+| 1.0.0   | [PR#414](https://github.com/bbc/psammead/pull/414) Add support for different scripts typographies |
+| 0.3.4   | [PR#407](https://github.com/bbc/psammead/pull/407) Organise dependencies properly |
+| 0.3.3   | [PR#392](https://github.com/bbc/psammead/pull/392) Increase padding-bottom from 16px to 24px |
+| 0.3.2   | [PR#398](https://github.com/bbc/psammead/pull/398) Update dependencies. Using `rem` for media queries. |
+| 0.3.1   | [PR#323](https://github.com/bbc/psammead/pull/323) Update storybook badge url |
+| 0.3.0   | [PR#318](https://github.com/BBC/psammead/pull/318) Update to new font face and family. |
+| 0.2.2   | [PR#323](https://github.com/BBC/psammead/pull/323) Update readme storybook badge |
 | 0.2.1   | [PR#310](https://github.com/BBC/psammead/pull/310) Update package-lock file. |
 | 0.2.0   | [PR#291](https://github.com/BBC/psammead/pull/291) Update appearance to reflect UX designs. |
 | 0.1.8   | [PR#264](https://github.com/BBC/psammead/pull/264) Resolving package-lock issues. |