+++ packages/utilities/gel-foundations/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 | Version | Description |
 |---------|-------------|
+| 1.2.0   | [PR#431](https://github.com/bbc/psammead/pull/431) Add custom scripts propTypes |
+| 1.1.0   | [PR#424](https://github.com/bbc/psammead/pull/424) Add Snyk badge to readme |
+| 1.0.0   | [PR#353](https://github.com/BBC/psammead/pull/353) Add support for different scripts typographies |
+| 0.3.0   | [PR#360](https://github.com/BBC/psammead/pull/360) Move to using rems for breakpoints |
+| 0.2.5   | [PR#349](https://github.com/BBC/psammead/pull/349) Add GEL Group A breakpoint |
+| 0.2.4   | [PR#323](https://github.com/BBC/psammead/pull/323) Update readme storybook badge |
 | 0.2.3   | [PR#264](https://github.com/BBC/psammead/pull/319) Resolving package-lock issues. |
 | 0.2.2   | [PR#304](https://github.com/BBC-News/psammead/pull/304) Add font family definitions, which include GEL-defined font fallbacks. |
 | 0.2.1   | [PR#303](https://github.com/BBC-News/psammead/pull/303) Bump to ensure correct build is published |