+++ packages/utilities/psammead-assets/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 | Version | Description |
 |---------|-------------|
-| 0.1.5    | [PR#264](https://github.com/BBC/psammead/pull/319) Resolving package-lock issues. |
+| 0.1.8   | [PR#444](https://github.com/bbc/psammead/pull/444) Add service SVG's |
+| 0.1.7   | [PR#424](https://github.com/bbc/psammead/pull/424) Add Snyk badge to readme |
+| 0.1.6   | [PR#323](https://github.com/BBC/psammead/pull/323) Update readme storybook badge |
+| 0.1.5   | [PR#264](https://github.com/BBC/psammead/pull/319) Resolving package-lock issues. |
 | 0.1.4   | [PR#224](https://github.com/BBC-News/psammead/pull/224) Add tests for the exported values, coverage 100% :tada: |
 | 0.1.3   | [PR#212](https://github.com/BBC-News/psammead/pull/212) Update package description and README. |
 | 0.1.2   | [PR#173](https://github.com/BBC-News/psammead/pull/173) Update PRs welcome link |